Refactor final command display to a separate component for improved readability and maintainability
