+++ apps/electron/src/renderer/components/mcp/server/server-details/ServerDetailsAdvancedSheet.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { MCPServer } from "@mcp_router/shared";
 import { useTranslation } from "react-i18next";
-import { Settings2, Check, RefreshCw, Terminal, Info, FileText, Plus, Trash, AlertTriangle } from "lucide-react";
+import { Settings2, Check, RefreshCw, Info, FileText, Plus, Trash, Terminal } from "lucide-react";
 import {
   Sheet,
   SheetContent,
@@ -13,9 +13,9 @@ import {
 import { Button } from "@mcp_router/ui";
 import { Input } from "@mcp_router/ui";
 import { Label } from "@mcp_router/ui";
-import { ScrollArea } from "@mcp_router/ui";
 import { Badge } from "@mcp_router/ui";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@mcp_router/ui";
+import FinalCommandDisplay from "./FinalCommandDisplay";
 import ServerDetailsRemote from "./ServerDetailsRemote";
 import ServerDetailsEnvironment from "./ServerDetailsEnvironment";
 import ServerDetailsAutoStart from "./ServerDetailsAutoStart";
@@ -56,35 +56,6 @@ const ServerDetailsAdvancedSheet: React.FC<ServerDetailsAdvancedSheetProps> = ({
     addEnvPair,
   } = useServerEditingStore();
   
-  // Function to substitute parameters in arguments
-  const getSubstitutedArgs = (
-    args: string[],
-    params: Record<string, string> = {},
-  ) => {
-    return args.map((arg) => {
-      // Check if the arg is a parameter reference like "{test}"
-      const paramMatch = arg.match(/^\{([^}]+)\}$/);
-      if (paramMatch && paramMatch[1]) {
-        const paramName = paramMatch[1];
-        // Use the input param value, fall back to default value
-        const paramValue =
-          params[paramName] || server.inputParams?.[paramName]?.default || arg;
-        return paramValue;
-      }
-      return arg;
-    });
-  };
-
-  // Get the final command string with args
-  const getFinalCommandString = () => {
-    if (!server.command) return "";
-
-    const command = server.command;
-    if (!server.args || server.args.length === 0) return command;
-
-    const substitutedArgs = getSubstitutedArgs(server.args, inputParamValues);
-    return `${command} ${substitutedArgs.join(" ")}`;
-  };
   
   // State for input parameters
   const [inputParamValues, setInputParamValues] = useState<Record<string, string>>({});
@@ -175,30 +146,10 @@ const ServerDetailsAdvancedSheet: React.FC<ServerDetailsAdvancedSheetProps> = ({
           <TabsContent value="general" className="space-y-6 mt-4">
             {/* Final Command Display */}
             {server.serverType === "local" ? (
-              <div className="space-y-3">
-                <div className="flex items-center gap-2">
-                  <Terminal className="h-4 w-4 text-muted-foreground" />
-                  <h3 className="text-sm font-medium text-primary">
-                    {t("serverDetails.finalCommand")}
-                  </h3>
-                </div>
-                <div className="pl-6">
-                  {server.command ? (
-                    <div className="bg-muted p-3 rounded-md border shadow-sm">
-                      <ScrollArea className="max-h-[150px]">
-                        <div className="whitespace-pre-wrap text-sm font-mono text-primary/90 break-all">
-                          {getFinalCommandString()}
-                        </div>
-                      </ScrollArea>
-                    </div>
-                  ) : (
-                    <div className="flex items-center gap-2 text-muted-foreground italic p-2">
-                      <Info className="h-4 w-4" />
-                      <span>{t("serverDetails.notConfigured")}</span>
-                    </div>
-                  )}
-                </div>
-              </div>
+              <FinalCommandDisplay
+                server={server}
+                inputParamValues={inputParamValues}
+              />
             ) : (
               <ServerDetailsRemote
                 server={server}
